Fix #9 Fallback to clock_gettime() for getnsecs()
I did a bit more than the subject says:

* Use the POSIX clock (clock_gettime()) for all Linux builds.

* Start conversion to C99. Stop using C89 extensions like long long.

* Clean up code to my liking for all getnsecs/getusecs.

* Use short CDDL header (same as illumos).

+++ libmicro.h
@@ -1,26 +1,16 @@
 /*
- * CDDL HEADER START
+ * This file and its contents are supplied under the terms of the
+ * Common Development and Distribution License ("CDDL"), version 1.0.
+ * You may only use this file in accordance with the terms of version
+ * 1.0 of the CDDL.
  *
- * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
- *
- * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
- * or http://www.opensolaris.org/os/licensing.
- * See the License for the specific language governing permissions
- * and limitations under the License.
- *
- * When distributing Covered Code, include this CDDL HEADER in each
- * file and include the License file at usr/src/OPENSOLARIS.LICENSE.
- * If applicable, add the following below this CDDL HEADER, with the
- * fields enclosed by brackets "[]" replaced with your own identifying
- * information: Portions Copyright [yyyy] [name of copyright owner]
- *
- * CDDL HEADER END
+ * A full copy of the text of the CDDL should have accompanied this
+ * source.  A copy of the CDDL is also available via the Internet at
+ * http://www.illumos.org/license/CDDL.
  */
 
 /*
+ * Copyright 2015 Ryan Zezeski <ryan@zinascii.com>.
  * Copyright 2007 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
@@ -29,6 +19,7 @@
 #define	LIBMICRO_H
 
 #include <pthread.h>
+#include <stdint.h>
 
 #define	LIBMICRO_VERSION	"0.4.2"
 
@@ -185,8 +176,8 @@ extern size_t			lm_tsdsize;
 int 		getpindex();
 int 		gettindex();
 void 		*gettsd(int p, int t);
-long long 	getusecs();
-long long 	getnsecs();
+uint64_t 	getusecs();
+uint64_t 	getnsecs();
 int 		setfdlimit(int limit);
 long long 	sizetoll();
 int 		sizetoint();
